Generate Workspace using Gita (#47)
* Generate Workspace using Gita

This change adds the capability to generate a workspace directly from the known_good.json file. To achieve this, the known_good.json file is transformed into a configuration file for a git workspace management tool called gita, which is then used to check out all mentioned modules on the specific commit/branch that is given in known_good.json. In addition, the score_modules.MODULE.bazel is updated with local_path_overrides to the checked out modules. Hence, a bazel build will immediately work as expected.

# Conflicts:
#	.gitignore

# Conflicts:
#	.devcontainer/devcontainer.json

* also generate .gitmodules metadata

* Remove git submodules generation

This does not work, since submodules require also commit tracking in the top-level repository. This would by default pollute reference_integration, which is certainly not what we want.

* do not create a workspace by default

* Add VSCode Tasks and documentation

* Fix deprecation warning

* mark as executable

* clarify the devcontainer relation and which build targets should be working